+++ client.h
@@ -1058,6 +1058,7 @@ void CL_ClearProjectiles (void);
 void CL_ParsePacketEntities (qbool delta);
 void CL_SetSolidEntities (void);
 void CL_ParsePlayerinfo (void);
+void CL_StorePausePredictionLocations(void);
 
 
 void MVD_Interpolate(void);