+++ release-notes.md
@@ -75,6 +75,7 @@
 - Fixed bug causing 'crouch' step-smoothing to kick in when respawning at a (very) nearby spawn point (old bug)
 - Fixed bug causing "Unset entity number" message & program termination as entnum overwritten from local baselines after avoiding buffer overflow in previous frame (reported during sdCup2 on Discord)
 - Fixed bug causing player to suddenly rotate when game was unpaused as mouse movement during pause suddenly took effect (old bug)
+- Fixed bug causing other players to move around on screen when game is paused (old bug)
 
 ### Ruleset-related changes